refactor
the namespace wasn't matching the hud name
